backend/requirements.txt
# Core Dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.1.3
numpy>=1.26.0
scikit-learn>=1.3.2

# ML/Forecasting - Compatible with Python 3.12
tensorflow>=2.16.1
prophet>=1.1.5
joblib>=1.3.2

# LangGraph & AI
langgraph>=0.0.26
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0

# API & Data Collection
requests>=2.31.0
httpx>=0.25.2

# Visualization
matplotlib>=3.8.2
plotly>=5.18.0

# Database
sqlalchemy>=2.0.23

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.1
